Refactoring app.py and my.kv to split each class and faction and field. Created ./db, ./src/kv_files, ./src/common.
