+++ .gitignore
@@ -150,4 +150,4 @@ out
 gen
 
 # exclude db
-user_data.db
+db/user_data.db
