+++ my.kv
@@ -1,4 +1,8 @@
 
+#:include src/kv_files/main_scene.kv
+#:include src/kv_files/history_scene.kv
+#:include src/kv_files/setting_scene.kv
+
 Root:
     orientation: 'vertical'
 
@@ -36,150 +40,3 @@ Root:
         MainScene:
         HistoryScene:
         SettingScene:
-
-
-<MainScene>
-    name: "main"
-
-    total_saving: total_saving
-    store: store
-    balance: balance
-
-    FloatLayout:
-
-        BoxLayout:
-            orientation: 'vertical'
-            size_hint: 0.5, 0.5
-            pos_hint: {"center_x": .5, "center_y": .5}
-
-            MDLabel:
-                id: store
-                text: store.text
-                theme_text_color: 'Primary'
-
-            MDLabel:
-                id: balance
-                text: balance.text
-                theme_text_color: 'Primary'
-
-        BoxLayout:
-            size_hint: 0.5, 0.5
-            pos_hint: {"center_x": .75, "center_y": .5}
-
-            MDProgressBar:
-                orientation: 'vertical'
-                id: total_saving
-                max: total_saving.max
-                value: total_saving.value if total_saving.value < total_saving.max else total_saving.max
-                color: app.theme_cls.accent_color
-
-
-<HistoryScene>
-    name: "deposit"
-
-    deposit: deposit
-    button_deposit: button_deposit
-    scroll: scroll
-
-    FloatLayout:
-
-        BoxLayout:
-            size_hint: 0.4, 0.2
-            pos_hint: {"center_x": .5, "y": .7}
-
-            MDTextField:
-                id: deposit
-                input_filter: "float"
-                hint_text: "Enter amount"
-                helper_text: "Only accept positive amount."
-                helper_text_mode: "on_error"
-                pos_hint: {"center_x": .5, "center_y": .5}
-
-            MDRoundFlatButton:
-                id: button_deposit
-                text: "Deposit"
-                pos_hint: {"center_x": .5, "center_y": .5}
-                on_release: root.update_history()
-
-        BoxLayout:
-            size_hint: 0.8, 0.5
-            pos_hint: {"center_x": .5, "y": .1}
-            ScrollView:
-                do_scroll_x: False
-                do_scroll_y: True
-                MDList:
-                    id: scroll
-
-
-<ListItemWithCheckbox>:
-    id: id_widget
-    IconRightWidget:
-        on_release:
-            if checkbox.active:\
-            app.root.screen_manager.screens[1].delete_history_log(id_widget)
-        icon: root.icon
-
-    LeftCheckbox:
-        id: checkbox
-
-<SettingScene>
-    name: "setting"
-
-    objective: objective
-    button_objective: button_objective
-    button_reset: button_reset
-
-    FloatLayout:
-
-        BoxLayout:
-            size_hint: 0.8, 0.2
-            pos_hint: {"center_x": .5, "center_y": .8}
-
-            MDLabel:
-                text: 'Your Goal'
-                halign: "left"
-                theme_text_color: 'Primary'
-
-            MDTextField:
-                id: objective
-                input_filter: "float"
-                hint_text: "Enter amount"
-                helper_text: "Only accept positive amount."
-                helper_text_mode: "on_error"
-                pos_hint: {"center_x": .5, "center_y": .5}
-
-            MDRoundFlatButton:
-                id: button_objective
-                text: "Submit"
-                pos_hint: {"center_x": .5, "center_y": .5}
-                on_release: root.submit_objective()
-
-        BoxLayout:
-            size_hint: 0.8, 0.2
-            pos_hint: {"center_x": .5, "center_y": .5}
-            MDLabel:
-                text: 'Reset'
-                halign: "left"
-                theme_text_color: 'Primary'
-            MDLabel:
-                text: 'reset all history and your goal.'
-                halign: "left"
-                theme_text_color: 'Primary'
-            MDFlatButton:
-                id: button_reset
-                text: "Submit"
-                pos_hint: {"center_x": .5, "center_y": .5}
-                on_release: root.show_alert_dialog()
-
-
-
-        BoxLayout:
-            size_hint: 0.4, 0.2
-            pos_hint: {"center_x": .5, "center_y": .1}
-            MDIcon:
-                icon: 'theme-light-dark'
-                theme_text_color: 'Primary'
-                pos_hint: {"center_y": .5}
-            MDSwitch:
-                pos_hint: {"center_y": .5}
-                on_active: app.toggle_theme(*args)
