+++ src/common/config.py
@@ -3,3 +3,4 @@
 msg_objective = 'your goal $ '
 
 msg_balance = 'your balance $ '
+
