+++ src/common/const.py
@@ -1,0 +1,2 @@
+
+DATABASE_FILENAME = './db/user_data.db'  # Do not change unless you are willing to move or change db file.
