+++ src/kv_files/main_scene.kv
@@ -1,0 +1,34 @@
+<MainScene>
+    name: "main"
+
+    total_saving: total_saving
+    store: store
+    balance: balance
+
+    FloatLayout:
+
+        BoxLayout:
+            orientation: 'vertical'
+            size_hint: 0.5, 0.5
+            pos_hint: {"center_x": .5, "center_y": .5}
+
+            MDLabel:
+                id: store
+                text: store.text
+                theme_text_color: 'Primary'
+
+            MDLabel:
+                id: balance
+                text: balance.text
+                theme_text_color: 'Primary'
+
+        BoxLayout:
+            size_hint: 0.5, 0.5
+            pos_hint: {"center_x": .75, "center_y": .5}
+
+            MDProgressBar:
+                orientation: 'vertical'
+                id: total_saving
+                max: total_saving.max
+                value: total_saving.value if total_saving.value < total_saving.max else total_saving.max
+                color: app.theme_cls.accent_color
